+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - program: include usdc oracle ([#397](https://github.com/drift-labs/protocol-v2/pull/397))
+- ts-sdk: add addAllUsers to DriftClient
 
 ### Fixes